Changed message position and the color of the panel.
In the code on which this project was based a JLabel is used to pass a
message to the user when the game is completed. This Jlabel was
previously added to the South position of the JPanel (the JPanel being
set to a layout of type GridLayout) which represents the frame and
contains the 9 cells. This update changed the message to appear in the
North position of the panel's gridlayout.

In this update I also changed the color of the border of the JPanel
object panel to blue.